move routes and functions
